chore: extract components into separate files
